requirements.txt
kafka-python==2.0.2
flask==3.0.0
flask-cors==4.0.0
numpy==1.24.3
pymysql==1.1.0
python-dateutil==2.8.2